fix test and simplifying
