+++ data/beaconrestapi/src/integration-test/resources/tech/pegasys/teku/beaconrestapi/beacon/schema/BeaconStateBellatrix.json
@@ -151,7 +151,7 @@
       "$ref" : "#/components/schemas/SyncCommittee"
     },
     "latest_execution_payload_header" : {
-      "$ref" : "#/components/schemas/ExecutionPayloadHeader"
+      "$ref" : "#/components/schemas/ExecutionPayloadHeaderBellatrix"
     }
   }
 }
